\name{PILSNERSim-package}
\alias{PILSNERSim-package}
\alias{PILSNERSim}
\docType{package}
\title{
\packageTitle{PILSNERSim}
}
\description{
\packageDescription{PILSNERSim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PILSNERSim}
\packageIndices{PILSNERSim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PILSNERSim}

Maintainer: \packageMaintainer{PILSNERSim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
